Question backend/requirements.txt
# FastAPI core
fastapi
uvicorn

# LangChain ecosystem
langchain
langchain-core
langchain-community
langchain-groq
langchain-chroma
langchain-huggingface
langchain-google-genai
langchain-unstructured
langchain-exa
langchain-tavily
huggingface_hub[hf_xet]

newspaper3k
lxml-html-clean  # fixed name from `lxml_html_clean`
beautifulsoup4
docx2txt
json5

# Utilities
python-dotenv
httpx
pydantic
ddgs  # DuckDuckGo search wrapper
firecrawl-py
pymupdf
python-multipart